requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Chroma==0.2.0
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
confection==0.1.5
contourpy==1.3.1
curl_cffi==0.10.0
cycler==0.12.1
cymem==2.0.11
dataloader==2.0
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
keras-nightly==3.9.0.dev2025032003
kiwisolver==1.4.8
langchain==0.3.22
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langcodes==3.5.0
langsmith==0.3.22
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.44.0
lxml==5.3.2
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
multitasking==0.0.11
murmurhash==1.0.12
namex==0.0.8
narwhals==1.29.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.0.2
od==2.0.2
openai==1.68.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.16
packaging==24.2
panda==0.3.1
pandas==2.2.3
pandas-datareader==0.10.0
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.50
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytube==15.0.0
pytubefix==8.12.2
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.47.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
stack-data==0.6.3
streamlit==1.43.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_keras==0.1
termcolor==2.5.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.2.18
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
ttkbootstrap==1.10.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
utils==1.0.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.58
zstandard==0.23.0
